% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_covid19_controlled_vocab.R
\name{get_covid19_controlled_vocab}
\alias{get_covid19_controlled_vocab}
\title{Get Information About the Elements of the covid19R controlled vocabularies}
\usage{
get_covid19_controlled_vocab(vocab)
}
\arguments{
\item{vocab}{name of a column from a dataset pulled from \code{\link{get_covid19_dataset}} or a data type listed in \code{\link{get_covid19_data_info}}}
}
\value{
a tibble with information about approved controlled vocabulary terms and their description
}
\description{
Many columns of the tidy format for covid19R data use a controlled vocabulary to define standardized types. While you can read more about these in our documentation, this pulls the controlled vocabulary and description for a given vocabulary type.
}
\examples{

get_covid19_controlled_vocab("data_type")

}
